compare_ranges: rework support for comma-separated multi ranges
We have support for comma separated ranges in one CommitRange object.
That works because git_log_table() supports it. But it is in a very bad
relations with the fact that parse_range doesn't support comma, and
concept of CommitRange.top, CommitRange.base fields conflicts with
multirange too.

Let's support comma directly in CommitRange:

 - create list of Commit object in __init__, parsing comma
 - don't expose .git_range field. Instead we expose .commits field,
   so Table class can use it instead. It's better: don't call git-log
   subprocess twice.
 - be explicit in fact that .base and .top are available only for
   simple range, when no comma in the input definition

Signed-off-by: Vladimir Sementsov-Ogievskiy <vsementsov@virtuozzo.com>