rework parse_range: treat one commit as one commit
Historically we treat single commit as the range of the whole history
of this commit, like git-log does.

But that is in bad connection with
1. multirange feature (we want simple syntax of several commits
   separated by comma, which means exactly these several commits)
2. our with to simply compare two exact commits (git change-difftool
   is broken and will be removed soon)

So, let's treat one commit as one commit.

To support old feature of whole history (which is actually limited by
default base, as searching commits through the whole history is a time
waste), add new syntax: ..<commit>

We should also improve handling default_base in git-change-rebase:
use CommitRange class directly to parse and add new helpful exception.

Signed-off-by: Vladimir Sementsov-Ogievskiy <vsementsov@virtuozzo.com>