check_rebase_meta: improve grouping syntax
We have <tag>: syntax for describing groups. Tags started with "drop"
means drop-group.

We want to move to more explicit things:

1. Features.

Ideally, every commit is related to some feature, and I dream that
"Feature: <...>" tag would be in each downstream commit. While that is
not true, we want to define features in metadata file. So, the syntax
would be:

%feature <feature>
<commit1>
<commit2>
%end

 - instead of old syntax

<feature>:
<commit1>
<commit2>

Note, that new syntax has explicit "%end".

2. Drop-groups

For declaring commits to be dropped, drop-groups are introduced:

%drop
<commit1>
<commit2>
%end

or

%drop short-reason
<commit1>
<commit2>
%end

 - instead of old syntax

drop-short-reason:
<commit1>
<commit2>

Old syntax is still supported as deprecated.

Signed-off-by: Vladimir Sementsov-Ogievskiy <vsementsov@yandex-team.ru>