commit_ranges: include upstreaming/drop info into new/up columns
"upstreaming" is what about to upstream the commit. It may be a note
like "not-needed" or issue key in issue tracker. And it doesn't make
sense if commit already exist in upstream. So, insert this information
into git-range column called "up" if it exist.

"drop" means we want to drop this commit.

When we are just examine the downstream branch and compare it with
upstream, "drop" means that we are not going the commit (we want to
drop it). So this information is good for "up" column.

When we examine the rebase process, "drop" is more about porting the
commit to the new base point. So, we try to insert this information to
the git-range column called "new" if it exist.

So, "drop" info is removed from META column.

Signed-off-by: Vladimir Sementsov-Ogievskiy <vsementsov@yandex-team.ru>